refactor: reorganize the file locations
