--- packages/core/src/cli.ts
@@ -1,1 +1,0 @@
-export { runCLI } from './cli/index';
