+++ packages/core/src/config.ts
@@ -13,27 +13,6 @@ import {
   logger,
 } from './utils';
 
-export type RstestConfigAsyncFn = () => Promise<RstestConfig>;
-
-export type RstestConfigSyncFn = () => RstestConfig;
-
-export type RstestConfigExport =
-  | RstestConfig
-  | RstestConfigSyncFn
-  | RstestConfigAsyncFn;
-
-/**
- * This function helps you to autocomplete configuration types.
- * It accepts a Rsbuild config object, or a function that returns a config.
- */
-export function defineConfig(config: RstestConfig): RstestConfig;
-export function defineConfig(config: RstestConfigSyncFn): RstestConfigSyncFn;
-export function defineConfig(config: RstestConfigAsyncFn): RstestConfigAsyncFn;
-export function defineConfig(config: RstestConfigExport): RstestConfigExport;
-export function defineConfig(config: RstestConfigExport) {
-  return config;
-}
-
 const findConfig = (basePath: string): string | undefined => {
   return DEFAULT_CONFIG_EXTENSIONS.map((ext) => basePath + ext).find(
     fs.existsSync,
