--- packages/core/src/index.ts
@@ -1,7 +1,0 @@
-export { defineConfig } from './config';
-
-export * from './api/public';
-
-// TODO: Move below to @rstest/node exports point.
-export { createRstest } from './core';
-export type { RstestConfig, RstestCommand } from './types';
