+++ packages/core/src/node.ts
@@ -1,3 +1,3 @@
-export { defineConfig } from './config';
+export { defineConfig } from './public';
 export { createRstest } from './core';
 export type { RstestConfig, Reporter, RstestCommand } from './types';
