+++ packages/core/src/runtime/runner/runner.ts
@@ -1,5 +1,5 @@
 import { GLOBAL_EXPECT, getState, setState } from '@vitest/expect';
-import { getSnapshotClient } from '../api/snapshot';
+import { getSnapshotClient } from '../../api/snapshot';
 import type {
   AfterEachListener,
   BeforeEachListener,
@@ -12,9 +12,9 @@ import type {
   TestResultStatus,
   TestSuite,
   WorkerState,
-} from '../types';
-import { ROOT_SUITE_NAME } from '../utils';
-import { formatTestError } from '../utils/runtime';
+} from '../../types';
+import { ROOT_SUITE_NAME } from '../../utils';
+import { formatTestError } from '../util';
 
 const getTestStatus = (results: TestResult[]): TestResultStatus => {
   if (results.length === 0) {
