+++ packages/core/src/runtime/worker/loadModule.ts
@@ -1,7 +1,7 @@
 import { createRequire as createNativeRequire } from 'node:module';
 import vm from 'node:vm';
 import path from 'pathe';
-import { logger } from '../utils/logger';
+import { logger } from '../../utils/logger';
 
 const isRelativePath = (p: string) => /^\.\.?\//.test(p);
